[Added]: Handling if query data contains newlines, [Refactor]: Separated pipeline and component aspects
