+++ qanary_commons/src/main/java/eu/wdaqua/qanary/explainability/annotations/LoggingAspectPipeline.java
@@ -1,21 +1,15 @@
 package eu.wdaqua.qanary.explainability.annotations;
 
 
-import eu.wdaqua.qanary.commons.QanaryMessage;
-import eu.wdaqua.qanary.commons.triplestoreconnectors.QanaryTripleStoreConnector;
-import eu.wdaqua.qanary.commons.triplestoreconnectors.QanaryTripleStoreConnectorQanaryInternal;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
 import org.aspectj.lang.annotation.Pointcut;
-import org.springframework.stereotype.Service;
 
-import java.net.URI;
 import java.net.URISyntaxException;
 
 @Aspect
-@Service
-public class LoggingAspectPipeline {
+public class LoggingAspectPipeline extends LoggingAspect{
 
     ////////////////// CONFIG FOR PIPELINE
     /*
@@ -25,13 +19,10 @@ public class LoggingAspectPipeline {
      * QanaryTripleStoreEndpoint: Fetched from
      */
 
-    private URI processGraph;
-    private QanaryTripleStoreConnector qanaryTripleStoreConnector;
-
     @Pointcut(
             "execution(public org.springframework.http.ResponseEntity<?> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.startquestionansweringwithtextquestion(..)) || " +
-                    "execution(public org.springframework.http.ResponseEntity<eu.wdaqua.qanary.message.QanaryQuestionAnsweringRun> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.startquestionansweringwithtextquestion(...) || " +
-                    "execution(public org.springframework.http.ResponseEntity<eu.wdaqua.qanary.message.QanaryQuestionAnsweringRun> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.startquestionansweringwithtextquestionThroughJson(...) || " +
+                    "execution(public org.springframework.http.ResponseEntity<eu.wdaqua.qanary.message.QanaryQuestionAnsweringRun> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.startquestionansweringwithtextquestion(..)) || " +
+                    "execution(public org.springframework.http.ResponseEntity<eu.wdaqua.qanary.message.QanaryQuestionAnsweringRun> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.startquestionansweringwithtextquestionThroughJson(..)) || " +
                     "execution(public org.springframework.http.ResponseEntity<?> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.createQuestionAnswering(..)) || " +
                     "execution(public org.springframework.http.ResponseEntity<?> eu.wdaqua.qanary.web.QanaryQuestionAnsweringController.createQuestionAnsweringFull(..))"
     ) public void startProcessForPipeline() {};
@@ -41,11 +32,7 @@ public class LoggingAspectPipeline {
 
     @Before(value = "setTriplestoreAndGraphForPipeline()") // TODO: Can a parent class be established providing those methods for setting QTSC and graph?
     public void setTriplestoreAndGraphForPipeline(JoinPoint joinPoint) throws URISyntaxException {
-        QanaryMessage qanaryMessage = (QanaryMessage) joinPoint.getArgs()[2]; // 3rd parameter is of type QanaryMessage
-        this.processGraph = qanaryMessage.getInGraph();
-        if(this.qanaryTripleStoreConnector == null) {
-            this.qanaryTripleStoreConnector = new QanaryTripleStoreConnectorQanaryInternal(qanaryMessage.getEndpoint(), "null"); // TODO: Application name
-        }
+
     }
 
 }
