+++ qanary_commons/src/test/java/qa/commons/LoggingAspectComponentTest.java
